More moving stuff around, prepping for ES6
