+++ bin/sicksync.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('../index')();
+require('../src/index')();
