+++ src/commands/remove-project.js
@@ -1,5 +1,5 @@
 var _ = require('lodash'),
-    projectHelper = require('../lib/project-helper');
+    projectHelper = require('../project-helper');
 
 module.exports = function removeProjectCommand(program, config) {
     program
