+++ test/big-sync.mspec.js
@@ -2,7 +2,7 @@ var expect = require('chai').expect,
     _ = require('lodash'),
     rewire = require('rewire'),
     sinon = require('sinon'),
-    bigSync = rewire('../lib/big-sync');
+    bigSync = rewire('../src/big-sync');
 
 var mockConfig = {
     username: 'joel',
