+++ test/local/fs-helper.mspec.js
@@ -2,7 +2,7 @@ var _ = require('lodash'),
     expect = require('chai').expect,
     rewire = require('rewire'),
     sinon = require('sinon'),
-    FSHelper = rewire('../../lib/local/fs-helper'),
+    FSHelper = rewire('../../src/local/fs-helper'),
     config = {
         sourceLocation: 'my/home/',
         destinationLocation: 'my/remote/box/',
